Fixed bugs in symbols and integrated into rest of project
