+++ src/index.html
@@ -46,11 +46,11 @@
         const id = ComponentManager.createEmptyComponent("score-component");
         //const id = ComponentManager.createEmptyComponent("text-component");
         SelectionManager.select(id);
-        ComponentManager.toggleComponentSymbolEnabledState("component-0", "snare");
-        ComponentManager.toggleComponentSymbolEnabledState("component-0", "kick");
+        ComponentManager.toggleComponentDrumEnabledState("component-0", "snare");
+        ComponentManager.toggleComponentDrumEnabledState("component-0", "kick");
         
         attachEditor(document.getElementById("editor-pane"));
-        attachRendered(document.getElementById("component-container"));
+        //attachRendered(document.getElementById("component-container"));
